pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "azure-cost-optimizer"
version = "0.1.0"
description = "Multi-agent Azure cost analysis and optimization tool powered by Claude"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "azure-identity>=1.17.0",
    "azure-mgmt-costmanagement>=4.0.0",
    "azure-mgmt-compute>=32.0.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-mgmt-monitor>=6.0.0",
    "azure-mgmt-advisor>=9.0.0",
    "azure-mgmt-network>=25.0.0",
    "azure-monitor-query>=1.3.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[project.scripts]
azure-cost-optimizer = "main:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
